+++ src/lib/components/EvmMethodChip.tsx
@@ -1,0 +1,15 @@
+import { Tag } from "@chakra-ui/react";
+import type { TagProps } from "@chakra-ui/react";
+
+import { getEvmMethod } from "lib/utils";
+
+interface EvmMethodChipProps {
+  txInput: string;
+  width: TagProps["width"];
+}
+
+export const EvmMethodChip = ({ txInput, width }: EvmMethodChipProps) => (
+  <Tag width={width} height="17px" variant="gray" justifyContent="center">
+    {getEvmMethod(txInput)}
+  </Tag>
+);