+++ src/lib/pages/tx-details/index.tsx
@@ -9,7 +9,7 @@ import { Loading } from "lib/components/Loading";
 import PageContainer from "lib/components/PageContainer";
 import { CelatoneSeo } from "lib/components/Seo";
 import { EmptyState } from "lib/components/state/EmptyState";
-import { useTxData } from "lib/services/tx";
+import { useEvmTxHashByCosmosTxHash, useTxData } from "lib/services/tx";
 import { getFirstQueryParam, truncate } from "lib/utils";
 
 import { TxHeader, TxInfo, TxInfoMobile } from "./components";
@@ -20,6 +20,8 @@ const TxDetails = () => {
   const hashParam = getFirstQueryParam(router.query.txHash);
   const isMobile = useMobile();
   const { data, isLoading } = useTxData(hashParam);
+  const { data: relatedEvmTxHash, isFetching: isRelatedEvmTxFetching } =
+    useEvmTxHashByCosmosTxHash(hashParam);
 
   useEffect(() => {
     if (router.isReady && !isLoading) {
@@ -35,7 +37,8 @@ const TxDetails = () => {
     }
   }, [router.isReady, data, isLoading]);
 
-  if (isLoading || !hashParam) return <Loading withBorder />;
+  if (!hashParam || isLoading || isRelatedEvmTxFetching)
+    return <Loading withBorder />;
 
   return (
     <PageContainer>
@@ -50,9 +53,9 @@ const TxDetails = () => {
         <>
           <TxHeader mt={2} txData={data} />
           {isMobile && <TxInfoMobile txData={data} />}
-          <Flex my={{ base: 0, md: 12 }} justify="space-between">
+          <Flex my={{ base: 0, md: 12 }} gap={4} justify="space-between">
             {!isMobile && <TxInfo txData={data} />}
-            <MessageSection txData={data} />
+            <MessageSection txData={data} relatedEvmTxHash={relatedEvmTxHash} />
           </Flex>
         </>
       ) : (