+++ src/lib/services/types/tx.ts
@@ -12,6 +12,10 @@ import {
   MsgFurtherAction,
   zBechAddr,
   zCoin,
+  zHex,
+  zHexAddr20,
+  zHexBig,
+  zHexBool,
   zMessageResponse,
   zPagination,
   zPubkeyMulti,
@@ -35,7 +39,10 @@ import { zAny } from "./protobuf";
 
 const zModeInfoSingle = z.object({
   single: z.object({
-    mode: z.custom<SignMode>((val) => SignMode[val as keyof typeof SignMode]),
+    mode: z.union([
+      z.custom<SignMode>((val) => SignMode[val as keyof typeof SignMode]),
+      z.literal(9999), // minievm sign mode,
+    ]),
   }),
 });
 
@@ -162,15 +169,16 @@ export const zTxsResponseItemFromLcd =
       type: msg["@type"],
     }));
 
-    const { isIbc, isOpinit } = messages.reduce(
+    const { isIbc, isOpinit, isEvm } = messages.reduce(
       (acc, msg, idx) => {
         const current = getTxBadges(msg.type, logs[idx]);
         return {
           isIbc: acc.isIbc || current.isIbc,
           isOpinit: acc.isOpinit || current.isOpinit,
+          isEvm: acc.isEvm || current.isEvm,
         };
       },
-      { isIbc: false, isOpinit: false }
+      { isIbc: false, isOpinit: false, isEvm: false }
     );
 
     return {
@@ -183,6 +191,7 @@ export const zTxsResponseItemFromLcd =
       success: val.code === 0,
       isIbc,
       isOpinit,
+      isEvm,
       events: val.events,
       // TODO: implement below later
       actionMsgType: ActionMsgType.OTHER_ACTION_MSG,
@@ -255,6 +264,7 @@ const zBaseTxsResponseItem = z.object({
   is_send: z.boolean(),
   // initia
   is_opinit: z.boolean().optional(),
+  is_evm: z.boolean().optional(),
   // wasm
   is_clear_admin: z.boolean().optional(),
   is_execute: z.boolean().optional(),
@@ -291,6 +301,7 @@ export const zTxsResponseItem = zBaseTxsResponseItem.transform<Transaction>(
     furtherAction: MsgFurtherAction.NONE,
     isIbc: val.is_ibc,
     isOpinit: val.is_opinit ?? false,
+    isEvm: val.is_evm ?? false,
     isInstantiate: val.is_instantiate ?? false,
   })
 );
@@ -344,6 +355,7 @@ const zAccountTxsResponseItem = zBaseTxsResponseItem
     ),
     isIbc: val.is_ibc,
     isOpinit: val.is_opinit ?? false,
+    isEvm: val.is_evm ?? false,
     isInstantiate: val.is_instantiate ?? false,
   }));
 
@@ -368,3 +380,56 @@ export const zBlockTxsResponseSequencer = z.object({
   txs: z.array(zTxsResponseItemFromLcd),
   pagination: zPagination,
 });
+
+// ---------------- JSON RPC ----------------
+export const zEvmTxHashByCosmosTxHashJsonRpc = z.string().transform((val) =>
+  val !== "0x0000000000000000000000000000000000000000000000000000000000000000" // if no related evm tx
+    ? val
+    : null
+);
+
+export const zTxJsonRpc = z.object({
+  blockHash: z.string(),
+  blockNumber: zHexBig,
+  from: zHexAddr20,
+  gas: zHexBig,
+  gasPrice: zHexBig,
+  maxFeePerGas: zHexBig,
+  maxPriorityFeePerGas: zHexBig,
+  hash: z.string(),
+  input: z.string(),
+  nonce: zHexBig,
+  to: zHexAddr20.nullable(),
+  transactionIndex: zHexBig,
+  value: zHexBig,
+  type: z.string(), // TODO: convert to enum later
+  chainId: zHexBig,
+  v: z.string(),
+  r: z.string(),
+  s: z.string(),
+  yParity: z.string(),
+});
+
+export const zTxReceiptJsonRpc = z.object({
+  blockHash: z.string(),
+  blockNumber: zHexBig,
+  contractAddress: zHexAddr20.nullable(),
+  cumulativeGasUsed: zHexBig,
+  effectiveGasPrice: zHexBig,
+  from: zHexAddr20,
+  gasUsed: zHexBig,
+  logs: z.object({}).passthrough().array(),
+  logsBloom: zHex,
+  status: zHexBool,
+  to: zHexAddr20.nullable(),
+  transactionHash: z.string(),
+  transactionIndex: zHexBig,
+  type: z.string(), // TODO: convert to enum later
+});
+
+export const zTxDataJsonRpc = z
+  .tuple([zTxJsonRpc, zTxReceiptJsonRpc])
+  .transform(([tx, txReceipt]) => ({ tx, txReceipt }));
+export type TxDataJsonRpc = z.infer<typeof zTxDataJsonRpc>;
+
+export type TxDataWithTimeStampJsonRpc = TxDataJsonRpc & { timestamp: Date };