+++ src/lib/types/chainConfig.ts
@@ -46,6 +46,16 @@ const zMoveConfig = z.union([
   }),
 ]);
 
+const zEvmConfig = z.union([
+  z.object({
+    enabled: z.literal(true),
+    jsonRpc: z.string(),
+  }),
+  z.object({
+    enabled: z.literal(false),
+  }),
+]);
+
 const zPoolConfig = z.union([
   z.object({
     enabled: z.literal(true),
@@ -210,6 +220,7 @@ export const zChainConfig = z
       faucet: zFaucetConfig,
       gov: zGovConfig,
       move: zMoveConfig,
+      evm: zEvmConfig,
       nft: zNftConfig,
       pool: zPoolConfig,
       publicProject: zPublicProjectConfig,